+++ core/src/main/java/org/keycloak/representations/config/BaseRealmConfig.java
@@ -1,0 +1,64 @@
+package org.keycloak.representations.config;
+
+import org.codehaus.jackson.annotate.JsonProperty;
+import org.codehaus.jackson.annotate.JsonPropertyOrder;
+
+/**
+ * Common Realm Configuration
+ *
+ * @author <a href="mailto:bill@burkecentral.com">Bill Burke</a>
+ * @version $Revision: 1 $
+ */
+@JsonPropertyOrder({"realm", "realm-public-key", "auth-url", "code-url", "ssl-not-required"})
+public class BaseRealmConfig {
+    @JsonProperty("realm")
+    protected String realm;
+    @JsonProperty("realm-public-key")
+    protected String realmKey;
+    @JsonProperty("auth-url")
+    protected String authUrl;
+    @JsonProperty("code-url")
+    protected String codeUrl;
+    @JsonProperty("ssl-not-required")
+    protected boolean sslNotRequired;
+
+    public boolean isSslNotRequired() {
+        return sslNotRequired;
+    }
+
+    public void setSslNotRequired(boolean sslNotRequired) {
+        this.sslNotRequired = sslNotRequired;
+    }
+
+    public String getRealm() {
+        return realm;
+    }
+
+    public void setRealm(String realm) {
+        this.realm = realm;
+    }
+
+    public String getRealmKey() {
+        return realmKey;
+    }
+
+    public void setRealmKey(String realmKey) {
+        this.realmKey = realmKey;
+    }
+
+    public String getAuthUrl() {
+        return authUrl;
+    }
+
+    public void setAuthUrl(String authUrl) {
+        this.authUrl = authUrl;
+    }
+
+    public String getCodeUrl() {
+        return codeUrl;
+    }
+
+    public void setCodeUrl(String codeUrl) {
+        this.codeUrl = codeUrl;
+    }
+}
