--- core/src/main/java/org/keycloak/representations/idm/ApplicationInstallationRepresentation.java
@@ -1,78 +1,0 @@
-package org.keycloak.representations.idm;
-
-import org.codehaus.jackson.annotate.JsonProperty;
-
-import java.util.Map;
-
-/**
- * @author <a href="mailto:sthorger@redhat.com">Stian Thorgersen</a>
- */
-public class ApplicationInstallationRepresentation {
-    protected String realm;
-    protected String resource;
-    @JsonProperty("realm-public-key")
-    protected String realmPublicKey;
-    @JsonProperty("auth-url")
-    protected String authUrl;
-    @JsonProperty("code-url")
-    protected String codeUrl;
-    @JsonProperty("ssl-not-required")
-    protected boolean sslNotRequired;
-    protected Map<String, String> credentials;
-
-    public String getRealm() {
-        return realm;
-    }
-
-    public void setRealm(String realm) {
-        this.realm = realm;
-    }
-
-    public String getResource() {
-        return resource;
-    }
-
-    public void setResource(String resource) {
-        this.resource = resource;
-    }
-
-    public String getRealmPublicKey() {
-        return realmPublicKey;
-    }
-
-    public void setRealmPublicKey(String realmPublicKey) {
-        this.realmPublicKey = realmPublicKey;
-    }
-
-    public String getAuthUrl() {
-        return authUrl;
-    }
-
-    public void setAuthUrl(String authUrl) {
-        this.authUrl = authUrl;
-    }
-
-    public String getCodeUrl() {
-        return codeUrl;
-    }
-
-    public void setCodeUrl(String codeUrl) {
-        this.codeUrl = codeUrl;
-    }
-
-    public boolean isSslNotRequired() {
-        return sslNotRequired;
-    }
-
-    public void setSslNotRequired(boolean sslNotRequired) {
-        this.sslNotRequired = sslNotRequired;
-    }
-
-    public Map<String, String> getCredentials() {
-        return credentials;
-    }
-
-    public void setCredentials(Map<String, String> credentials) {
-        this.credentials = credentials;
-    }
-}
