+++ core/src/main/java/org/keycloak/util/JsonSerialization.java
@@ -1,6 +1,7 @@
 package org.keycloak.util;
 
 import org.codehaus.jackson.map.ObjectMapper;
+import org.codehaus.jackson.map.SerializationConfig;
 import org.codehaus.jackson.map.annotate.JsonSerialize;
 
 import java.io.IOException;
@@ -14,10 +15,14 @@ import java.io.InputStream;
  */
 public class JsonSerialization {
     public static final ObjectMapper mapper = new ObjectMapper();
+    public static final ObjectMapper prettyMapper = new ObjectMapper();
 
     static {
         mapper.setSerializationInclusion(JsonSerialize.Inclusion.NON_DEFAULT);
         mapper.setSerializationInclusion(JsonSerialize.Inclusion.NON_NULL);
+        prettyMapper.enable(SerializationConfig.Feature.INDENT_OUTPUT);
+        prettyMapper.setSerializationInclusion(JsonSerialize.Inclusion.NON_DEFAULT);
+        prettyMapper.setSerializationInclusion(JsonSerialize.Inclusion.NON_NULL);
     }
 
     public static String writeValueAsString(Object obj) throws IOException {
