--- integration/adapter-core/src/main/java/org/keycloak/adapters/config/AdapterConfig.java
@@ -1,243 +1,0 @@
-package org.keycloak.adapters.config;
-
-import org.codehaus.jackson.annotate.JsonProperty;
-import org.codehaus.jackson.annotate.JsonPropertyOrder;
-
-import java.util.HashMap;
-import java.util.Map;
-
-/**
- * @author <a href="mailto:bill@burkecentral.com">Bill Burke</a>
- * @version $Revision: 1 $
- */
-@JsonPropertyOrder({"realm-url", "realm", "resource", "realm-public-key", "admin-role", "auth-url", "code-url", "allow-any-hostname", "disable-trust-manager", "truststore", "truststore-password", "client-id", "client-credentials"})
-public class AdapterConfig
-{
-    @JsonProperty("realm-url")
-    protected String realmUrl;
-    @JsonProperty("realm")
-    protected String realm;
-    @JsonProperty("resource")
-    protected String resource;
-    @JsonProperty("realm-public-key")
-    protected String realmKey;
-    @JsonProperty("auth-url")
-    protected String authUrl;
-    @JsonProperty("code-url")
-    protected String codeUrl;
-    @JsonProperty("use-resource-role-mappings")
-    protected boolean useResourceRoleMappings;
-
-    @JsonProperty("ssl-not-required")
-    protected boolean sslNotRequired;
-    @JsonProperty("allow-any-hostname")
-    protected boolean allowAnyHostname;
-    @JsonProperty("disable-trust-manager")
-    protected boolean disableTrustManager;
-    @JsonProperty("truststore")
-    protected String truststore;
-    @JsonProperty("truststore-password")
-    protected String truststorePassword;
-    @JsonProperty("client-keystore")
-    protected String clientKeystore;
-    @JsonProperty("client-keystore-password")
-    protected String clientKeystorePassword;
-    @JsonProperty("client-key-password")
-    protected String clientKeyPassword;
-    @JsonProperty("credentials")
-    protected Map<String, String> credentials = new HashMap<String, String>();
-    @JsonProperty("connection-pool-size")
-    protected int connectionPoolSize;
-    @JsonProperty("enable-cors")
-    protected boolean cors;
-    @JsonProperty("cors-max-age")
-    protected int corsMaxAge = -1;
-    @JsonProperty("cors-allowed-headers")
-    protected String corsAllowedHeaders;
-    @JsonProperty("cors-allowed-methods")
-    protected String corsAllowedMethods;
-    @JsonProperty("expose-token")
-    protected boolean exposeToken;
-    @JsonProperty("bearer-only")
-    protected boolean bearerOnly;
-
-    public boolean isUseResourceRoleMappings() {
-        return useResourceRoleMappings;
-    }
-
-    public void setUseResourceRoleMappings(boolean useResourceRoleMappings) {
-        this.useResourceRoleMappings = useResourceRoleMappings;
-    }
-
-    public boolean isSslNotRequired() {
-        return sslNotRequired;
-    }
-
-    public void setSslNotRequired(boolean sslNotRequired) {
-        this.sslNotRequired = sslNotRequired;
-    }
-
-    public String getRealmUrl() {
-        return realmUrl;
-    }
-
-    public void setRealmUrl(String realmUrl) {
-        this.realmUrl = realmUrl;
-    }
-
-    public String getRealm() {
-        return realm;
-    }
-
-    public void setRealm(String realm) {
-        this.realm = realm;
-    }
-
-    public String getResource() {
-        return resource;
-    }
-
-    public void setResource(String resource) {
-        this.resource = resource;
-    }
-
-    public String getRealmKey() {
-        return realmKey;
-    }
-
-    public void setRealmKey(String realmKey) {
-        this.realmKey = realmKey;
-    }
-
-    public String getAuthUrl() {
-        return authUrl;
-    }
-
-    public void setAuthUrl(String authUrl) {
-        this.authUrl = authUrl;
-    }
-
-    public String getCodeUrl() {
-        return codeUrl;
-    }
-
-    public void setCodeUrl(String codeUrl) {
-        this.codeUrl = codeUrl;
-    }
-
-    public boolean isAllowAnyHostname() {
-        return allowAnyHostname;
-    }
-
-    public void setAllowAnyHostname(boolean allowAnyHostname) {
-        this.allowAnyHostname = allowAnyHostname;
-    }
-
-    public boolean isDisableTrustManager() {
-        return disableTrustManager;
-    }
-
-    public void setDisableTrustManager(boolean disableTrustManager) {
-        this.disableTrustManager = disableTrustManager;
-    }
-
-    public String getTruststore() {
-        return truststore;
-    }
-
-    public void setTruststore(String truststore) {
-        this.truststore = truststore;
-    }
-
-    public String getTruststorePassword() {
-        return truststorePassword;
-    }
-
-    public void setTruststorePassword(String truststorePassword) {
-        this.truststorePassword = truststorePassword;
-    }
-
-    public Map<String, String> getCredentials() {
-        return credentials;
-    }
-
-    public String getClientKeystore() {
-        return clientKeystore;
-    }
-
-    public void setClientKeystore(String clientKeystore) {
-        this.clientKeystore = clientKeystore;
-    }
-
-    public String getClientKeystorePassword() {
-        return clientKeystorePassword;
-    }
-
-    public void setClientKeystorePassword(String clientKeystorePassword) {
-        this.clientKeystorePassword = clientKeystorePassword;
-    }
-
-    public String getClientKeyPassword() {
-        return clientKeyPassword;
-    }
-
-    public void setClientKeyPassword(String clientKeyPassword) {
-        this.clientKeyPassword = clientKeyPassword;
-    }
-
-    public int getConnectionPoolSize() {
-        return connectionPoolSize;
-    }
-
-    public void setConnectionPoolSize(int connectionPoolSize) {
-        this.connectionPoolSize = connectionPoolSize;
-    }
-
-   public boolean isCors() {
-        return cors;
-    }
-
-    public void setCors(boolean cors) {
-        this.cors = cors;
-    }
-
-    public int getCorsMaxAge() {
-        return corsMaxAge;
-    }
-
-    public void setCorsMaxAge(int corsMaxAge) {
-        this.corsMaxAge = corsMaxAge;
-    }
-
-    public String getCorsAllowedHeaders() {
-        return corsAllowedHeaders;
-    }
-
-    public void setCorsAllowedHeaders(String corsAllowedHeaders) {
-        this.corsAllowedHeaders = corsAllowedHeaders;
-    }
-
-    public String getCorsAllowedMethods() {
-        return corsAllowedMethods;
-    }
-
-    public void setCorsAllowedMethods(String corsAllowedMethods) {
-        this.corsAllowedMethods = corsAllowedMethods;
-    }
-
-    public boolean isExposeToken() {
-        return exposeToken;
-    }
-
-    public void setExposeToken(boolean exposeToken) {
-        this.exposeToken = exposeToken;
-    }
-
-    public boolean isBearerOnly() {
-        return bearerOnly;
-    }
-
-    public void setBearerOnly(boolean bearerOnly) {
-        this.bearerOnly = bearerOnly;
-    }
-}
